complete the custom uitableview cell where it holds the repo item data and building all of it components using auto layout and without interface builder start the task created inside githubService to fetch the api add an extension to Double where we can format double into friendly format

+++ githubrepos/Models/Item.swift
@@ -12,7 +12,7 @@ struct Item: Decodable {
     let id: Int
     let name: String
     let description: String
-    let stargazersCount: Int
+    let stargazersCount: Double
     let owner: Owner
     
     enum CodingKeys: String, CodingKey{
